circle

+++ src/main/java/figures/Figure.java
@@ -1,8 +1,8 @@
 package figures;
 
 public interface Figure {
-    int area();
-    int perimeter();
+    double area();
+    double perimeter();
     Figure move(int x, int y);
     boolean isInside(Point a);
 }
